fix career_member_users controller
