+++ app/views/web/admin/career_member_users/_careers.html.slim
