+++ app/views/web/admin/career_member_users/_search_form.html.slim
@@ -1,8 +1,8 @@
 .p-3.mb-3.bg-light
-  = search_form_for q, default_filter_form_options(url: admin_career_users_path) do |f|
+  = search_form_for q, default_filter_form_options(url: url_for) do |f|
     = f.input :last_name_or_first_name_cont, placeholder: "#{han('user', :last_name)}/#{han('user', :first_name)}", label: false
     = f.input :email_cont, placeholder: han('user', :email), label: false
-    = f.input :careers_slug_eq, placeholder: han('career', :name), label: false, as: :select_with_search, collection: @careers.map { |career| [career.name, career.slug] }
+    = f.input :careers_slug_eq, prompt: t('.careers'), label: false, as: :select_with_search, collection: @careers.map { |career| [career.name, career.slug] }
     .col.d-flex.align-self-end
       = f.button :submit, t('submit'), class: 'btn-primary me-2 flex-grow-1'
       = link_to t('reset'), url_for, class: 'btn btn-outline-primary'
