+++ app/views/web/admin/career_member_users/new.html.slim
@@ -1,6 +1,6 @@
 - content_for :header do
   = t('.add_user')
 
-= render 'web/admin/career_users/menu'
+= render 'menu'
 
 = render 'form', career_member: @career_member, careers: @careers
