--- app/views/web/admin/career_users/_menu.html.slim
@@ -1,5 +1,0 @@
-ul.nav.nav-tabs.mb-4
-  = menu_item t('.list'), admin_career_users_path
-  = menu_item t('.new'), new_admin_career_member_user_path
-  = menu_item t('.users_finished_career'), finished_admin_career_users_path
-  = menu_item t('.users_archived_career'), archived_admin_career_users_path
